oneplus2: Enable CNE and DPM
This is a combination of 5 commits.
The first commit's message is:
DPM: Enable DPM to help with current drain

Change-Id: I5fcbb6259eaa843cbcf057e23d23bfc7ddbffd3e

This is the 2nd commit message:

sambar: add missing DPM blobs

Change-Id: Ib76d5ea74874b342371a70caab4d964cc7be4888

This is the 3rd commit message:

CNE: DPM: Enable CNE and enable LD_PRELOAD of libNimsWrap.so

issue-id: SAMBAR-785

Change-Id: I7f63b01af676b9c9dff7f2dc9e4fbd61df92d067

This is the 4th commit message:

sambar: update proprietary files list

Change-Id: I55bdfbdd40e79334ef954a8c28d50a6249dbd853

This is the 5th commit message:

sambar: Move cne and dpm config to blobs

It's proprietary and is outdated.

Change-Id: I269e6a546054444007dbe944b6391083198d04ae

Change-Id: I815226cf3ecbf175b8b47515f8e0cd2639fcf10e