oneplus2: Fix sched_boost parameters
Commit 1cdbbf74 is not correct. The sched_boost parameters we set in
init.qcom.rc is executed on init to speed up boot, but that commit
moved it to boot, which is after init. What's worse, the parameter is
set after init power scripts are executed, so it will leave sched_boost
always on, which is expected to be set back to zero.

Fix it by revert that commit. The sepolicy part is left as we may make
use of that.

Change-Id: Ib8d10dbd5101e4f3c36a7c6c9a32e4f93adde226